#9 Naming refactor - directory
