+++ lib/use_cases/use_case_result.dart
